requirements.txt
backports.functools-lru-cache==1.5
cheroot==6.5.8
CherryPy==18.2.0
cycler==0.10.0
decorator==4.4.0
imageio==2.5.0
jaraco.functools==2.0
kiwisolver==1.1.0
matplotlib==3.1.1
more-itertools==7.2.0
networkx==2.3
numpy==1.17.1
Pillow==6.1.0
pkg-resources==0.0.0
portend==2.5
protobuf==3.9.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
tempora==1.14.1
zc.lockfile==2.0
